% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Rand index}
\usage{
adjusted_rand_index(x, y)
}
\arguments{
\item{x}{A vector of labels.}

\item{y}{A vector of labels with the same length as \code{x}.}
}
\value{
A numeric adjusted Rand index.
}
\description{
Computes the adjusted Rand index between two partitions.
}
