% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{available_lpari_examples}
\alias{available_lpari_examples}
\title{List bundled example data sets}
\usage{
available_lpari_examples()
}
\value{
A data frame describing the examples bundled with the package.
}
\description{
List bundled example data sets
}
