% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lpari_coefficients}
\alias{lpari_coefficients}
\title{Study 1 LPA-RI coefficients}
\format{
A data frame with columns:
\describe{
\item{term}{Model term.}
\item{estimate}{Logistic-regression coefficient.}
}
}
\usage{
lpari_coefficients
}
\description{
A coefficient table used by \code{\link[=lpari_score_candidates]{lpari_score_candidates()}} to compute the LPA-RI
linear predictor. These coefficients were trained in Study 1 simulations to
predict exact recovery of the true number of profiles.
}
\keyword{datasets}
