% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{lpari_feature_names}
\alias{lpari_feature_names}
\title{Names of the predictors used by LPA-RI}
\usage{
lpari_feature_names()
}
\value{
A character vector with the feature names expected by
\code{\link[=lpari_score_candidates]{lpari_score_candidates()}}.
}
\description{
Names of the predictors used by LPA-RI
}
